services/analytics-service/requirements.txt
Django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
requests==2.32.3
torch==2.5.1 --index-url https://download.pytorch.org/whl/cpu
numpy==1.26.4
scikit-learn==1.5.2
